requirements.txt
onnx==1.10.2
numpy==1.21.0
pillow==9.0.1
django==3.2.13
waitress==2.1.2
protobuf==3.20.0
whitenoise==5.3.0
onnxruntime==1.9.0
scikit-learn==1.0.2
opencv-python-headless
django-cors-headers==3.11.0